Moved the import hook to a private module
