+++ docs/versionhistory.rst
@@ -7,6 +7,8 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 
 - **BACKWARD INCOMPATIBLE** Removed the ``check_argument_types()`` and
   ``check_return_type()`` functions (use ``@typechecked`` instead)
+- **BACKWARD INCOMPATIBLE** Moved ``install_import_hook`` to be directly importable
+  from the ``typeguard`` module
 - Added support for PEP 604 union types (``X | Y``) on all Python versions (initial
   implementation contributed by supersergiy)
 - Added support for generic built-in collection types (``list[int]`` et al) on all
